+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+* Add `BakeRexWrappers` to every recipe-- gives an attribute for rex
 * Fix problem letter tag in multipart questions
 * Adds `meet-author` note in `neuroscience`
 * Add `lifestage-context` note to `nursing-internal`